pyproject.toml
[tool.poetry]
name = "pastypy"
version = "1.0.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.27.1"
pycryptodome = "^3.14.1"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
python-lsp-server = {extras = ["all"], version = "^1.3.3"}
black = "^22.1.0"
pytest-cov = "^3.0.0"
pytest-asyncio = "^0.18.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "strict"
filterwarnings = [
    'ignore::pytest.PytestUnraisableExceptionWarning'
]
addopts = "--cov=pastypy --cov-report=term-missing"
